@@ -8,16 +8,15 @@
 	<name>resources-service</name>
 	<version>VERSION</version>
 	<properties>
-		<java.version>11</java.version>
+		<java.version>21</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
-		<springdoc-version>1.8.0</springdoc-version>
-		<spring-boot.version>3.3.0</spring-boot.version>
+		<spring-boot.version>3.5.0</spring-boot.version>
 	</properties>
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.6</version>
+		<version>3.5.0</version>
 	</parent>
 	<build>
 		<sourceDirectory>src/main/java</sourceDirectory>
