@@ -3,19 +3,16 @@ package org.epos.api;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.epos.api.beans.MonitoringBean;
-import org.epos.api.core.MonitoringGeneration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import javax.servlet.http.HttpServletRequest;
-import java.util.HashMap;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Map;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-10-11T14:51:06.469Z[GMT]")
+@jakarta.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-10-11T14:51:06.469Z[GMT]")
 @RestController
 public class MonitoringApiController extends ApiController implements MonitoringApi {
 
